created Dataset abstraction
