+++ tidy3d/components/data/monitor_data.py
@@ -1,21 +1,16 @@
-# pylint:disable=too-many-lines
 """ Monitor Level Data, store the DataArrays associated with a single monitor."""
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Union, Dict, Tuple, Callable, List
+from typing import Union, Tuple
 import warnings
 import xarray as xr
 import numpy as np
 import pydantic as pd
 
-from .data_array import ScalarFieldDataArray, ScalarFieldTimeDataArray, ScalarModeFieldDataArray
-from .data_array import FluxTimeDataArray, FluxDataArray, ModeIndexDataArray, ModeAmpsDataArray
-from .data_array import Near2FarAngleDataArray, Near2FarCartesianDataArray, Near2FarKSpaceDataArray
-from .data_array import DataArray, DiffractionDataArray
-
-from ..base import TYPE_TAG_STR, Tidy3dBaseModel
-from ..types import Axis, Coordinate, Symmetry
+from .data_array import DiffractionDataArray
+from ..base import TYPE_TAG_STR
+from ..types import Coordinate, Symmetry
 from ..grid.grid import Grid
 from ..validators import enforce_monitor_fields_present, required_if_symmetry_present
 from ..monitor import MonitorType, FieldMonitor, FieldTimeMonitor, ModeSolverMonitor
@@ -23,11 +18,17 @@ from ..monitor import ModeMonitor, FluxMonitor, FluxTimeMonitor, PermittivityMon
 from ..monitor import Near2FarAngleMonitor, Near2FarCartesianMonitor, Near2FarKSpaceMonitor
 from ..monitor import DiffractionMonitor
 from ..medium import Medium
-from ...log import DataError, SetupError, log
-from ...constants import ETA_0, C_0, MICROMETER
+from ...log import SetupError, log
+from ...constants import ETA_0, C_0
+from .dataset import Dataset, AbstractFieldDataset
+from .dataset import FieldDataset, FieldTimeDataset, ModeSolverDataset, PermittivityDataset
+from .dataset import ModeDataset, FluxDataset, FluxTimeDataset
+from .dataset import AbstractNear2FarDataset, Near2FarKSpaceDataset
+from .dataset import Near2FarAngleDataset, Near2FarCartesianDataset
+from .dataset import DiffractionDataset
 
 
-class MonitorData(Tidy3dBaseModel, ABC):
+class MonitorData(Dataset, ABC):
     """Abstract base class of objects that store data pertaining to a single :class:`.monitor`."""
 
     monitor: MonitorType = pd.Field(
@@ -42,13 +43,8 @@ class MonitorData(Tidy3dBaseModel, ABC):
         """Return copy of self with symmetry applied."""
         return self.copy()
 
-    # pylint:disable=unused-argument
-    def normalize(self, source_spectrum_fn: Callable[[float], complex]) -> MonitorData:
-        """Return copy of self after normalization is applied using source spectrum function."""
-        return self.copy()
 
-
-class AbstractFieldData(MonitorData, ABC):
+class AbstractFieldData(MonitorData, AbstractFieldDataset, ABC):
     """Collection of scalar fields with some symmetry properties."""
 
     monitor: Union[FieldMonitor, FieldTimeMonitor, PermittivityMonitor, ModeSolverMonitor]
@@ -74,21 +70,6 @@ class AbstractFieldData(MonitorData, ABC):
 
     _require_sym_center = required_if_symmetry_present("symmetry_center")
     _require_grid_expanded = required_if_symmetry_present("grid_expanded")
-
-    @property
-    @abstractmethod
-    def field_components(self) -> Dict[str, DataArray]:
-        """Maps the field components to thier associated data."""
-
-    @property
-    @abstractmethod
-    def grid_locations(self) -> Dict[str, str]:
-        """Maps field components to the string key of their grid locations on the yee lattice."""
-
-    @property
-    @abstractmethod
-    def symmetry_eigenvalues(self) -> Dict[str, Callable[[Axis], float]]:
-        """Maps field components to their (positive) symmetry eigenvalues."""
 
     @property
     def symmetry_expanded_copy(self) -> AbstractFieldData:
@@ -146,95 +127,13 @@ class AbstractFieldData(MonitorData, ABC):
 
         return self.copy(update=new_fields)
 
-    def colocate(self, x=None, y=None, z=None) -> xr.Dataset:
-        """colocate all of the data at a set of x, y, z coordinates.
 
-        Parameters
-        ----------
-        x : Optional[array-like] = None
-            x coordinates of locations.
-            If not supplied, does not try to colocate on this dimension.
-        y : Optional[array-like] = None
-            y coordinates of locations.
-            If not supplied, does not try to colocate on this dimension.
-        z : Optional[array-like] = None
-            z coordinates of locations.
-            If not supplied, does not try to colocate on this dimension.
-
-        Returns
-        -------
-        xr.Dataset
-            Dataset containing all fields at the same spatial locations.
-            For more details refer to `xarray's Documentaton <https://tinyurl.com/cyca3krz>`_.
-
-        Note
-        ----
-        For many operations (such as flux calculations and plotting),
-        it is important that the fields are colocated at the same spatial locations.
-        Be sure to apply this method to your field data in those cases.
-        """
-
-        # convert supplied coordinates to array and assign string mapping to them
-        supplied_coord_map = {k: np.array(v) for k, v in zip("xyz", (x, y, z)) if v is not None}
-
-        # dict of data arrays to combine in dataset and return
-        centered_fields = {}
-
-        # loop through field components
-        for field_name, field_data in self.field_components.items():
-
-            # loop through x, y, z dimensions and raise an error if only one element along dim
-            for coord_name, coords_supplied in supplied_coord_map.items():
-                coord_data = field_data.coords[coord_name]
-                if coord_data.size == 1:
-                    raise DataError(
-                        f"colocate given {coord_name}={coords_supplied}, but "
-                        f"data only has one coordinate at {coord_name}={coord_data.values[0]}. "
-                        "Therefore, can't colocate along this dimension. "
-                        f"supply {coord_name}=None to skip it."
-                    )
-
-            centered_fields[field_name] = field_data.interp(
-                **supplied_coord_map, kwargs={"bounds_error": True}
-            )
-
-        # combine all centered fields in a dataset
-        return xr.Dataset(centered_fields)
-
-
-class ElectromagneticFieldData(AbstractFieldData, ABC):
-    """Stores a collection of E and H fields with x, y, z components."""
-
-    @property
-    def field_components(self) -> Dict[str, DataArray]:
-        """Maps the field components to thier associated data."""
-        # pylint:disable=no-member
-        return {field: getattr(self, field) for field in self.monitor.fields}
-
-    @property
-    def grid_locations(self) -> Dict[str, str]:
-        """Maps field components to the string key of their grid locations on the yee lattice."""
-        return dict(Ex="Ex", Ey="Ey", Ez="Ez", Hx="Hx", Hy="Hy", Hz="Hz")
-
-    @property
-    def symmetry_eigenvalues(self) -> Dict[str, Callable[[Axis], float]]:
-        """Maps field components to their (positive) symmetry eigenvalues."""
-
-        return dict(
-            Ex=lambda dim: -1 if (dim == 0) else +1,
-            Ey=lambda dim: -1 if (dim == 1) else +1,
-            Ez=lambda dim: -1 if (dim == 2) else +1,
-            Hx=lambda dim: +1 if (dim == 0) else -1,
-            Hy=lambda dim: +1 if (dim == 1) else -1,
-            Hz=lambda dim: +1 if (dim == 2) else -1,
-        )
-
-
-class FieldData(ElectromagneticFieldData):
+class FieldData(FieldDataset, AbstractFieldData):
     """Data associated with a :class:`.FieldMonitor`: scalar components of E and H fields.
 
     Example
     -------
+    >>> from tidy3d import ScalarFieldDataArray
     >>> x = [-1,1]
     >>> y = [-2,0,2]
     >>> z = [-3,-1,1,3]
@@ -247,54 +146,15 @@ class FieldData(ElectromagneticFieldData):
 
     monitor: FieldMonitor
 
-    Ex: ScalarFieldDataArray = pd.Field(
-        None,
-        title="Ex",
-        description="Spatial distribution of the x-component of the electric field.",
-    )
-    Ey: ScalarFieldDataArray = pd.Field(
-        None,
-        title="Ey",
-        description="Spatial distribution of the y-component of the electric field.",
-    )
-    Ez: ScalarFieldDataArray = pd.Field(
-        None,
-        title="Ez",
-        description="Spatial distribution of the z-component of the electric field.",
-    )
-    Hx: ScalarFieldDataArray = pd.Field(
-        None,
-        title="Hx",
-        description="Spatial distribution of the x-component of the magnetic field.",
-    )
-    Hy: ScalarFieldDataArray = pd.Field(
-        None,
-        title="Hy",
-        description="Spatial distribution of the y-component of the magnetic field.",
-    )
-    Hz: ScalarFieldDataArray = pd.Field(
-        None,
-        title="Hz",
-        description="Spatial distribution of the z-component of the magnetic field.",
-    )
-
     _contains_monitor_fields = enforce_monitor_fields_present()
 
-    def normalize(self, source_spectrum_fn: Callable[[float], complex]) -> FieldData:
-        """Return copy of self after normalization is applied using source spectrum function."""
-        fields_norm = {}
-        for field_name, field_data in self.field_components.items():
-            src_amps = source_spectrum_fn(field_data.f)
-            fields_norm[field_name] = (field_data / src_amps).astype(field_data.dtype)
 
-        return self.copy(update=fields_norm)
-
-
-class FieldTimeData(ElectromagneticFieldData):
+class FieldTimeData(FieldTimeDataset, AbstractFieldData):
     """Data associated with a :class:`.FieldTimeMonitor`: scalar components of E and H fields.
 
     Example
     -------
+    >>> from tidy3d import ScalarFieldTimeDataArray
     >>> x = [-1,1]
     >>> y = [-2,0,2]
     >>> z = [-3,-1,1,3]
@@ -307,46 +167,16 @@ class FieldTimeData(ElectromagneticFieldData):
 
     monitor: FieldTimeMonitor
 
-    Ex: ScalarFieldTimeDataArray = pd.Field(
-        None,
-        title="Ex",
-        description="Spatial distribution of the x-component of the electric field.",
-    )
-    Ey: ScalarFieldTimeDataArray = pd.Field(
-        None,
-        title="Ey",
-        description="Spatial distribution of the y-component of the electric field.",
-    )
-    Ez: ScalarFieldTimeDataArray = pd.Field(
-        None,
-        title="Ez",
-        description="Spatial distribution of the z-component of the electric field.",
-    )
-    Hx: ScalarFieldTimeDataArray = pd.Field(
-        None,
-        title="Hx",
-        description="Spatial distribution of the x-component of the magnetic field.",
-    )
-    Hy: ScalarFieldTimeDataArray = pd.Field(
-        None,
-        title="Hy",
-        description="Spatial distribution of the y-component of the magnetic field.",
-    )
-    Hz: ScalarFieldTimeDataArray = pd.Field(
-        None,
-        title="Hz",
-        description="Spatial distribution of the z-component of the magnetic field.",
-    )
-
     _contains_monitor_fields = enforce_monitor_fields_present()
 
 
-class ModeSolverData(ElectromagneticFieldData):
+class ModeSolverData(ModeSolverDataset, AbstractFieldData):
     """Data associated with a :class:`.ModeSolverMonitor`: scalar components of E and H fields.
 
     Example
     -------
     >>> from tidy3d import ModeSpec
+    >>> from tidy3d import ScalarModeFieldDataArray, ModeIndexDataArray
     >>> x = [-1,1]
     >>> y = [0]
     >>> z = [-3,-1,1,3]
@@ -376,94 +206,13 @@ class ModeSolverData(ElectromagneticFieldData):
 
     monitor: ModeSolverMonitor
 
-    Ex: ScalarModeFieldDataArray = pd.Field(
-        ...,
-        title="Ex",
-        description="Spatial distribution of the x-component of the electric field of the mode.",
-    )
-    Ey: ScalarModeFieldDataArray = pd.Field(
-        ...,
-        title="Ey",
-        description="Spatial distribution of the y-component of the electric field of the mode.",
-    )
-    Ez: ScalarModeFieldDataArray = pd.Field(
-        ...,
-        title="Ez",
-        description="Spatial distribution of the z-component of the electric field of the mode.",
-    )
-    Hx: ScalarModeFieldDataArray = pd.Field(
-        ...,
-        title="Hx",
-        description="Spatial distribution of the x-component of the magnetic field of the mode.",
-    )
-    Hy: ScalarModeFieldDataArray = pd.Field(
-        ...,
-        title="Hy",
-        description="Spatial distribution of the y-component of the magnetic field of the mode.",
-    )
-    Hz: ScalarModeFieldDataArray = pd.Field(
-        ...,
-        title="Hz",
-        description="Spatial distribution of the z-component of the magnetic field of the mode.",
-    )
 
-    n_complex: ModeIndexDataArray = pd.Field(
-        ...,
-        title="Propagation Index",
-        description="Complex-valued effective propagation constants associated with the mode.",
-    )
-
-    @property
-    def field_components(self) -> Dict[str, DataArray]:
-        """Maps the field components to thier associated data."""
-        # pylint:disable=no-member
-        return {field: getattr(self, field) for field in ["Ex", "Ey", "Ez", "Hx", "Hy", "Hz"]}
-
-    @property
-    def n_eff(self):
-        """Real part of the propagation index."""
-        return self.n_complex.real
-
-    @property
-    def k_eff(self):
-        """Imaginary part of the propagation index."""
-        return self.n_complex.imag
-
-    def sel_mode_index(self, mode_index: pd.NonNegativeInt) -> FieldData:
-        """Return :class:`.FieldData` for the specificed mode index."""
-
-        fields = {}
-        for field_name, data in self.field_components.items():
-            data = data.sel(mode_index=mode_index)
-            coords = {dim: data.coords[dim] for dim in "xyzf"}
-            scalar_field = ScalarFieldDataArray(data.data, coords=coords)
-            fields[field_name] = scalar_field
-
-        monitor_dict = self.monitor.dict(exclude={TYPE_TAG_STR, "mode_spec"})
-        field_monitor = FieldMonitor(**monitor_dict)
-
-        return FieldData(
-            monitor=field_monitor,
-            symmetry=self.symmetry,
-            symmetry_center=self.symmetry_center,
-            grid_expanded=self.grid_expanded,
-            **fields,
-        )
-
-    def plot_field(self, *args, **kwargs):
-        """Warn user to use the :class:`.ModeSolver` ``plot_field`` function now."""
-        raise DeprecationWarning(
-            "The 'plot_field()' method was moved to the 'ModeSolver' object."
-            "Once the 'ModeSolver' is contructed, one may call '.plot_field()' on the object and "
-            "the modes will be computed and displayed with 'Simulation' overlay."
-        )
-
-
-class PermittivityData(AbstractFieldData):
+class PermittivityData(PermittivityDataset, AbstractFieldData):
     """Data for a :class:`.PermittivityMonitor`: diagonal components of the permittivity tensor.
 
     Example
     -------
+    >>> from tidy3d import ScalarFieldDataArray
     >>> x = [-1,1]
     >>> y = [-2,0,2]
     >>> z = [-3,-1,1,3]
@@ -476,44 +225,14 @@ class PermittivityData(AbstractFieldData):
 
     monitor: PermittivityMonitor
 
-    @property
-    def field_components(self) -> Dict[str, ScalarFieldDataArray]:
-        """Maps the field components to thier associated data."""
-        return dict(eps_xx=self.eps_xx, eps_yy=self.eps_yy, eps_zz=self.eps_zz)
 
-    @property
-    def grid_locations(self) -> Dict[str, str]:
-        """Maps field components to the string key of their grid locations on the yee lattice."""
-        return dict(eps_xx="Ex", eps_yy="Ey", eps_zz="Ez")
-
-    @property
-    def symmetry_eigenvalues(self) -> Dict[str, Callable[[Axis], float]]:
-        """Maps field components to their (positive) symmetry eigenvalues."""
-        return dict(eps_xx=None, eps_yy=None, eps_zz=None)
-
-    eps_xx: ScalarFieldDataArray = pd.Field(
-        ...,
-        title="Epsilon xx",
-        description="Spatial distribution of the x-component of the electric field.",
-    )
-    eps_yy: ScalarFieldDataArray = pd.Field(
-        ...,
-        title="Epsilon yy",
-        description="Spatial distribution of the y-component of the electric field.",
-    )
-    eps_zz: ScalarFieldDataArray = pd.Field(
-        ...,
-        title="Epsilon zz",
-        description="Spatial distribution of the z-component of the electric field.",
-    )
-
-
-class ModeData(MonitorData):
+class ModeData(MonitorData, ModeDataset):
     """Data associated with a :class:`.ModeMonitor`: modal amplitudes and propagation indices.
 
     Example
     -------
     >>> from tidy3d import ModeSpec
+    >>> from tidy3d import ModeAmpsDataArray, ModeIndexDataArray
     >>> direction = ["+", "-"]
     >>> f = [1e14, 2e14, 3e14]
     >>> mode_index = np.arange(5)
@@ -532,39 +251,13 @@ class ModeData(MonitorData):
 
     monitor: ModeMonitor
 
-    amps: ModeAmpsDataArray = pd.Field(
-        ..., title="Amplitudes", description="Complex-valued amplitudes associated with the mode."
-    )
 
-    n_complex: ModeIndexDataArray = pd.Field(
-        ...,
-        title="Propagation Index",
-        description="Complex-valued effective propagation constants associated with the mode.",
-    )
-
-    @property
-    def n_eff(self):
-        """Real part of the propagation index."""
-        return self.n_complex.real
-
-    @property
-    def k_eff(self):
-        """Imaginary part of the propagation index."""
-        return self.n_complex.imag
-
-    def normalize(self, source_spectrum_fn) -> ModeData:
-        """Return copy of self after normalization is applied using source spectrum function."""
-        if self.amps is None:
-            raise DataError("ModeData contains no amp data, can't normalize.")
-        source_freq_amps = source_spectrum_fn(self.amps.f)[None, :, None]
-        return self.copy(update={"amps": self.amps / source_freq_amps})
-
-
-class FluxData(MonitorData):
+class FluxData(MonitorData, FluxDataset):
     """Data associated with a :class:`.FluxMonitor`: flux data in the frequency-domain.
 
     Example
     -------
+    >>> from tidy3d import FluxDataArray
     >>> f = [2e14, 3e14]
     >>> coords = dict(f=f)
     >>> flux_data = FluxDataArray(np.random.random(2), coords=coords)
@@ -573,20 +266,14 @@ class FluxData(MonitorData):
     """
 
     monitor: FluxMonitor
-    flux: FluxDataArray
-
-    def normalize(self, source_spectrum_fn) -> FluxData:
-        """Return copy of self after normalization is applied using source spectrum function."""
-        source_freq_amps = source_spectrum_fn(self.flux.f)
-        source_power = abs(source_freq_amps) ** 2
-        return self.copy(update={"flux": self.flux / source_power})
 
 
-class FluxTimeData(MonitorData):
+class FluxTimeData(MonitorData, FluxTimeDataset):
     """Data associated with a :class:`.FluxTimeMonitor`: flux data in the time-domain.
 
     Example
     -------
+    >>> from tidy3d import FluxTimeDataArray
     >>> t = [0, 1e-12, 2e-12]
     >>> coords = dict(t=t)
     >>> flux_data = FluxTimeDataArray(np.random.random(3), coords=coords)
@@ -595,13 +282,9 @@ class FluxTimeData(MonitorData):
     """
 
     monitor: FluxTimeMonitor
-    flux: FluxTimeDataArray
 
 
-RADVECTYPE = Union[Near2FarAngleDataArray, Near2FarCartesianDataArray, Near2FarKSpaceDataArray]
-
-
-class AbstractNear2FarData(MonitorData, ABC):
+class AbstractNear2FarData(MonitorData, AbstractNear2FarDataset, ABC):
     """Collection of radiation vectors in the frequency domain."""
 
     monitor: Union[Near2FarAngleMonitor, Near2FarCartesianMonitor, Near2FarKSpaceMonitor] = None
@@ -612,80 +295,16 @@ class AbstractNear2FarData(MonitorData, ABC):
         description="Background medium in which to radiate near fields to far fields.",
     )
 
-    Ntheta: RADVECTYPE = pd.Field(
-        ...,
-        title="Ntheta",
-        description="Spatial distribution of the theta-component of the N radiation vector.",
-    )
-    Nphi: RADVECTYPE = pd.Field(
-        ...,
-        title="Nphi",
-        description="Spatial distribution of phi-component of the N radiation vector.",
-    )
-    Ltheta: RADVECTYPE = pd.Field(
-        ...,
-        title="Ltheta",
-        description="Spatial distribution of theta-component of the L radiation vector.",
-    )
-    Lphi: RADVECTYPE = pd.Field(
-        ...,
-        title="Lphi",
-        description="Spatial distribution of phi-component of the L radiation vector.",
-    )
-
-    @property
-    def field_components(self) -> Dict[str, DataArray]:
-        """Maps the field components to thier associated data."""
-        return dict(
-            Ntheta=self.Ntheta,
-            Nphi=self.Nphi,
-            Ltheta=self.Ltheta,
-            Lphi=self.Lphi,
-        )
-
-    @property
-    def f(self) -> np.ndarray:
-        """Frequencies."""
-        return self.Ntheta.f.values
-
-    @property
-    def coords(self) -> Dict[str, np.ndarray]:
-        """Coordinates of the radiation vectors contained."""
-        return self.Ntheta.coords
-
-    @property
-    def dims(self) -> Tuple[str, ...]:
-        """Dimensions of the radiation vectors contained."""
-        return self.Ntheta.__slots__
-
-    def make_data_array(self, data: np.ndarray) -> xr.DataArray:
-        """Make an xr.DataArray with data and same coords and dims as radiation vectors of self."""
-        return xr.DataArray(data=data, coords=self.coords, dims=self.dims)
-
-    def make_dataset(self, keys: Tuple[str, ...], vals: Tuple[np.ndarray, ...]) -> xr.Dataset:
-        """Make an xr.Dataset with keys and data with same coords and dims as radiation vectors."""
-        data_arrays = tuple(map(self.make_data_array, vals))
-        return xr.Dataset(dict(zip(keys, data_arrays)))
-
-    def normalize(self, source_spectrum_fn: Callable[[float], complex]) -> AbstractNear2FarData:
-        """Return copy of self after normalization is applied using source spectrum function."""
-        fields_norm = {}
-        for field_name, field_data in self.field_components.items():
-            src_amps = source_spectrum_fn(field_data.f)
-            fields_norm[field_name] = field_data / src_amps
-
-        return self.copy(update=fields_norm)
-
-    @property
-    def nk(self) -> Tuple[float, float]:
-        """Returns the real and imaginary parts of the background medium's refractive index."""
-        return self.medium.nk_model(frequency=self.f)
-
     @staticmethod
     def propagation_factor(medium: Medium, frequency: float) -> complex:
         """Complex valued wavenumber associated with a frequency."""
         index_n, index_k = medium.nk_model(frequency=frequency)
         return (2 * np.pi * frequency / C_0) * (index_n + 1j * index_k)
+
+    @property
+    def nk(self) -> Tuple[float, float]:
+        """Returns the real and imaginary parts of the background medium's refractive index."""
+        return self.medium.nk_model(frequency=self.f)
 
     @property
     def k(self) -> complex:
@@ -783,15 +402,16 @@ class AbstractNear2FarData(MonitorData, ABC):
         """
 
 
-class Near2FarAngleData(AbstractNear2FarData):
+class Near2FarAngleData(AbstractNear2FarData, Near2FarAngleDataset):
     """Data associated with a :class:`.Near2FarAngleMonitor`: components of radiation vectors.
 
     Example
     -------
+    >>> from tidy3d import Near2FarAngleDataArray
     >>> f = np.linspace(1e14, 2e14, 10)
     >>> theta = np.linspace(0, np.pi, 10)
     >>> phi = np.linspace(0, 2*np.pi, 20)
-    >>> coords = dict(f=f, theta=theta, phi=phi)
+    >>> coords = dict(theta=theta, phi=phi, f=f)
     >>> values = (1+1j) * np.random.random((len(theta), len(phi), len(f)))
     >>> scalar_field = Near2FarAngleDataArray(values, coords=coords)
     >>> monitor = Near2FarAngleMonitor(
@@ -805,38 +425,7 @@ class Near2FarAngleData(AbstractNear2FarData):
 
     monitor: Near2FarAngleMonitor = None
 
-    Ntheta: Near2FarAngleDataArray = pd.Field(
-        ...,
-        title="Ntheta",
-        description="Spatial distribution of the theta-component of the N radiation vector.",
-    )
-    Nphi: Near2FarAngleDataArray = pd.Field(
-        ...,
-        title="Nphi",
-        description="Spatial distribution of phi-component of the N radiation vector.",
-    )
-    Ltheta: Near2FarAngleDataArray = pd.Field(
-        ...,
-        title="Ltheta",
-        description="Spatial distribution of theta-component of the L radiation vector.",
-    )
-    Lphi: Near2FarAngleDataArray = pd.Field(
-        ...,
-        title="Lphi",
-        description="Spatial distribution of phi-component of the L radiation vector.",
-    )
-
     _contains_monitor_fields = enforce_monitor_fields_present()
-
-    @property
-    def theta(self) -> np.ndarray:
-        """Polar angles."""
-        return self.Ntheta.theta.values
-
-    @property
-    def phi(self) -> np.ndarray:
-        """Azimuthal angles."""
-        return self.Ntheta.phi.values
 
     def fields(self, r: float = None) -> xr.Dataset:
         """Get fields in spherical coordinates relative to the monitor's local origin
@@ -901,15 +490,16 @@ class Near2FarAngleData(AbstractNear2FarData):
         return self.make_data_array(data=power_data)
 
 
-class Near2FarCartesianData(AbstractNear2FarData):
+class Near2FarCartesianData(AbstractNear2FarData, Near2FarCartesianDataset):
     """Data associated with a :class:`.Near2FarCartesianMonitor`: components of radiation vectors.
 
     Example
     -------
+    >>> from tidy3d import Near2FarCartesianDataArray
     >>> f = np.linspace(1e14, 2e14, 10)
     >>> x = np.linspace(0, 5, 10)
     >>> y = np.linspace(0, 10, 20)
-    >>> coords = dict(f=f, x=x, y=y)
+    >>> coords = dict(x=x, y=y, f=f)
     >>> values = (1+1j) * np.random.random((len(x), len(y), len(f)))
     >>> scalar_field = Near2FarCartesianDataArray(values, coords=coords)
     >>> monitor = Near2FarCartesianMonitor(
@@ -924,38 +514,7 @@ class Near2FarCartesianData(AbstractNear2FarData):
 
     monitor: Near2FarCartesianMonitor
 
-    Ntheta: Near2FarCartesianDataArray = pd.Field(
-        ...,
-        title="Ntheta",
-        description="Spatial distribution of the theta-component of the N radiation vector.",
-    )
-    Nphi: Near2FarCartesianDataArray = pd.Field(
-        ...,
-        title="Nphi",
-        description="Spatial distribution of the phi-component of the N radiation vector.",
-    )
-    Ltheta: Near2FarCartesianDataArray = pd.Field(
-        ...,
-        title="Ltheta",
-        description="Spatial distribution of the theta-component of the L radiation vector.",
-    )
-    Lphi: Near2FarCartesianDataArray = pd.Field(
-        ...,
-        title="Lphi",
-        description="Spatial distribution of the phi-component of the L radiation vector.",
-    )
-
     _contains_monitor_fields = enforce_monitor_fields_present()
-
-    @property
-    def x(self) -> np.ndarray:
-        """X positions."""
-        return self.Ntheta.x.values
-
-    @property
-    def y(self) -> np.ndarray:
-        """Y positions."""
-        return self.Ntheta.y.values
 
     @property
     def spherical_coords(self) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
@@ -1030,15 +589,16 @@ class Near2FarCartesianData(AbstractNear2FarData):
         return self.make_data_array(data=power)
 
 
-class Near2FarKSpaceData(AbstractNear2FarData):
+class Near2FarKSpaceData(AbstractNear2FarData, Near2FarKSpaceDataset):
     """Data associated with a :class:`.Near2FarKSpaceMonitor`: components of radiation vectors.
 
     Example
     -------
+    >>> from tidy3d import Near2FarKSpaceDataArray
     >>> f = np.linspace(1e14, 2e14, 10)
     >>> ux = np.linspace(0, 5, 10)
     >>> uy = np.linspace(0, 10, 20)
-    >>> coords = dict(f=f, ux=ux, uy=uy)
+    >>> coords = dict(ux=ux, uy=uy, f=f)
     >>> values = (1+1j) * np.random.random((len(ux), len(uy), len(f)))
     >>> scalar_field = Near2FarKSpaceDataArray(values, coords=coords)
     >>> monitor = Near2FarKSpaceMonitor(
@@ -1052,38 +612,7 @@ class Near2FarKSpaceData(AbstractNear2FarData):
 
     monitor: Near2FarKSpaceMonitor = None
 
-    Ntheta: Near2FarKSpaceDataArray = pd.Field(
-        ...,
-        title="Ntheta",
-        description="Spatial distribution of the theta-component of the N radiation vector.",
-    )
-    Nphi: Near2FarKSpaceDataArray = pd.Field(
-        ...,
-        title="Nphi",
-        description="Spatial distribution of phi-component of the N radiation vector.",
-    )
-    Ltheta: Near2FarKSpaceDataArray = pd.Field(
-        ...,
-        title="Ltheta",
-        description="Spatial distribution of theta-component of the L radiation vector.",
-    )
-    Lphi: Near2FarKSpaceDataArray = pd.Field(
-        ...,
-        title="Lphi",
-        description="Spatial distribution of phi-component of the L radiation vector.",
-    )
-
     _contains_monitor_fields = enforce_monitor_fields_present()
-
-    @property
-    def ux(self) -> np.ndarray:
-        """reciprocal X positions."""
-        return self.Ntheta.ux.values
-
-    @property
-    def uy(self) -> np.ndarray:
-        """reciprocal Y positions."""
-        return self.Ntheta.uy.values
 
     # pylint:disable=too-many-locals
     def fields(self, r: float = None) -> xr.Dataset:
@@ -1138,24 +667,23 @@ class Near2FarKSpaceData(AbstractNear2FarData):
         return self.make_data_array(data=power_values)
 
 
-# pylint: disable=too-many-public-methods
-class DiffractionData(MonitorData):
-    """Data associated with a :class:`.DiffractionMonitor`:
-    complex components of diffracted far fields.
+class DiffractionData(MonitorData, DiffractionDataset):
+    """Data for a :class:`.DiffractionMonitor`: complex components of diffracted far fields.
 
     Example
     -------
+    >>> from tidy3d import DiffractionDataArray
     >>> f = np.linspace(1e14, 2e14, 10)
     >>> orders_x = list(range(-4, 5))
     >>> orders_y = list(range(-6, 7))
     >>> pol = ["s", "p"]
-    >>> coords = dict(f=f, orders_x=orders_x, orders_y=orders_y, polarization=pol)
+    >>> coords = dict(orders_x=orders_x, orders_y=orders_y, polarization=pol, f=f)
     >>> values = (1+1j) * np.random.random((len(orders_x), len(orders_y), len(pol), len(f)))
     >>> field = DiffractionDataArray(values, coords=coords)
     >>> monitor = DiffractionMonitor(
     ...     center=(1,2,3), size=(np.inf,np.inf,0), freqs=f, name='diffraction',
     ...     orders_x=orders_x, orders_y=orders_y
-    ...     )
+    ... )
     >>> data = DiffractionData(
     ...     monitor=monitor, L=field, N=field, sim_size=[1,1], bloch_vecs=[1,2]
     ... )
@@ -1163,74 +691,10 @@ class DiffractionData(MonitorData):
 
     monitor: DiffractionMonitor
 
-    sim_size: Tuple[float, float] = pd.Field(
-        ...,
-        title="Simulation size",
-        description="Simulation sizes in the local x and y directions.",
-        units=MICROMETER,
-    )
-
-    bloch_vecs: Tuple[float, float] = pd.Field(
-        ...,
-        title="Bloch vectors",
-        description="Bloch vectors along the local x and y directions in units of "
-        "``2 * pi / (simulation size along the respective dimension)``.",
-    )
-
-    L: DiffractionDataArray = pd.Field(
-        ...,
-        title="L",
-        description="Complex components of the far field radiation vectors associated with the "
-        "electric field for each polarization tangential to ``monitor.normal_axis``, "
-        "in a local Cartesian coordinate system whose z-axis is ``monitor.normal_axis``.",
-    )
-
-    N: DiffractionDataArray = pd.Field(
-        ...,
-        title="N",
-        description="Complex components of the far field radiation vectors associated with the "
-        "magnetic field for each polarization tangential to ``monitor.normal_axis``, "
-        "in a local Cartesian coordinate system whose z-axis is ``monitor.normal_axis``.",
-    )
-
-    @staticmethod
-    def shifted_orders(orders: Tuple[int], bloch_vec: float) -> np.ndarray:
-        """Diffraction orders shifted by the Bloch vector."""
-        return bloch_vec + np.atleast_1d(orders)
-
-    @property
-    def field_components(self) -> Dict[str, DataArray]:
-        """Maps the field components to thier associated data."""
-        return dict(L=self.L, N=self.N)
-
-    def normalize(self, source_spectrum_fn: Callable[[float], complex]) -> DiffractionData:
-        """Copy of self after normalization is applied using source spectrum function."""
-        fields_norm = {}
-        for field_name, field_data in self.field_components.items():
-            src_amps = source_spectrum_fn(field_data.f)
-            fields_norm[field_name] = field_data / src_amps
-
-        return self.copy(update=fields_norm)
-
     @property
     def medium(self) -> Medium:
         """Medium in which the near fields are recorded and propagated."""
         return self.monitor.medium
-
-    @property
-    def frequencies(self) -> np.ndarray:
-        """Frequencies associated with ``monitor``."""
-        return np.atleast_1d(self.L.f.values)
-
-    @property
-    def orders_x(self) -> np.ndarray:
-        """Allowed orders along x."""
-        return np.atleast_1d(self.L.orders_x.values)
-
-    @property
-    def orders_y(self) -> np.ndarray:
-        """Allowed orders along y."""
-        return np.atleast_1d(self.L.orders_y.values)
 
     @property
     def wavenumber(self) -> np.ndarray:
@@ -1239,44 +703,10 @@ class DiffractionData(MonitorData):
         return np.real(2.0 * np.pi * self.frequencies / C_0 * np.sqrt(epsilon))
 
     @property
-    def wavelength(self) -> np.ndarray:
-        """Wavelength at each frequency."""
-        return 2.0 * np.pi / self.wavenumber
-
-    @property
     def eta(self) -> np.ndarray:
         """Wavelength at each frequency."""
         epsilon = self.medium.eps_model(self.frequencies)
         return np.real(ETA_0 / np.sqrt(epsilon))
-
-    @property
-    def ux(self) -> np.ndarray:
-        """Normalized wave vector along x relative to ``local_origin`` and oriented
-        with respect to ``monitor.normal_dir``, normalized by the wave number in the
-        background medium."""
-        if self.sim_size[0] == 0:
-            return np.atleast_2d(0)
-        bloch_x = self.shifted_orders(self.orders_x, self.bloch_vecs[0])
-        return bloch_x[:, None] * 2.0 * np.pi / self.sim_size[0] / self.wavenumber[None, :]
-
-    @property
-    def uy(self) -> np.ndarray:
-        """Normalized wave vector along y relative to ``local_origin`` and oriented
-        with respect to ``monitor.normal_dir``, normalized by the wave number in the
-        background medium."""
-        if self.sim_size[1] == 0:
-            return np.atleast_2d(0)
-        bloch_y = self.shifted_orders(self.orders_y, self.bloch_vecs[1])
-        return bloch_y[:, None] * 2.0 * np.pi / self.sim_size[1] / self.wavenumber[None, :]
-
-    def _make_coords_for_pol(self, pol: Tuple[str, str]) -> Dict[str, Union[np.ndarray, List]]:
-        """Make a coordinates dictionary for a given pair of polarization names."""
-        coords = {}
-        coords["orders_x"] = np.atleast_1d(self.orders_x)
-        coords["orders_y"] = np.atleast_1d(self.orders_y)
-        coords["polarization"] = pol
-        coords["f"] = np.array(self.frequencies)
-        return coords
 
     @property
     def angles(self) -> Tuple[xr.DataArray]:
@@ -1288,9 +718,8 @@ class DiffractionData(MonitorData):
             warnings.filterwarnings(
                 "ignore", message="invalid value encountered in arcsin", category=RuntimeWarning
             )
-            thetas, phis = DiffractionMonitor.kspace_2_sph(
-                self.ux[:, None, :], self.uy[None, :, :], axis=2
-            )
+            ux, uy = self.reciprocal_vectors
+            thetas, phis = DiffractionMonitor.kspace_2_sph(ux[:, None, :], uy[None, :, :], axis=2)
 
         coords = self._make_coords_for_pol(["", ""])
         del coords["polarization"]
@@ -1298,19 +727,25 @@ class DiffractionData(MonitorData):
         phi_data = xr.DataArray(phis, coords=coords)
         return theta_data, phi_data
 
-    def car_2_sph(self, field: DiffractionDataArray) -> DiffractionDataArray:
-        """Transform field stored as a :class:`DiffractionDataArray` to spherical coordinates,
-        assuming they represent plane waves. Angles are restricted to within the light cone;
-        other values are set to `nan`."""
-        f_x = field.sel(polarization="x").values
-        f_y = field.sel(polarization="y").values
-        theta, phi = self.angles
-        f_phi = np.nan_to_num(-np.sin(phi) * f_x + np.cos(phi) * f_y)
-        f_theta = np.nan_to_num((f_x * np.cos(phi) + f_y * np.sin(phi)) * np.cos(theta))
+    @property
+    def amps(self) -> DiffractionDataArray:
+        """Complex power amplitude in each order for 's' and 'p' polarizations, normalized so that
+        the power carried by the wave of that order and polarization equals ``abs(amps)^2``.
+        """
+        cos_theta = np.cos(np.nan_to_num(self.angles[0]))
+        norm = 1.0 / np.sqrt(2.0 * ETA_0) / np.sqrt(cos_theta)
+        amp_theta = self.E_sph.sel(polarization="theta").values * norm
+        amp_phi = self.E_sph.sel(polarization="phi").values * norm
 
+        # stack the amplitudes in s- and p-components along a new polarization axis
         return DiffractionDataArray(
-            np.stack([f_theta, f_phi], axis=2), coords=self._make_coords_for_pol(["theta", "phi"])
+            np.stack([amp_phi, amp_theta], axis=2), coords=self._make_coords_for_pol(["s", "p"])
         )
+
+    @property
+    def power(self) -> xr.DataArray:
+        """Total power in each order, summed over both polarizations."""
+        return (np.abs(self.amps) ** 2).sum(dim="polarization")
 
     def sph_2_car(self, field: DiffractionDataArray) -> DiffractionDataArray:
         """Transform field stored as a :class:`DiffractionDataArray` to Cartesian coordinates,
@@ -1326,6 +761,20 @@ class DiffractionData(MonitorData):
             np.stack([f_x, f_y, f_z], axis=2), coords=self._make_coords_for_pol(["x", "y", "z"])
         )
 
+    def car_2_sph(self, field: DiffractionDataArray) -> DiffractionDataArray:
+        """Transform field stored as a :class:`DiffractionDataArray` to spherical coordinates,
+        assuming they represent plane waves. Angles are restricted to within the light cone;
+        other values are set to `nan`."""
+        f_x = field.sel(polarization="x").values
+        f_y = field.sel(polarization="y").values
+        theta, phi = self.angles
+        f_phi = np.nan_to_num(-np.sin(phi) * f_x + np.cos(phi) * f_y)
+        f_theta = np.nan_to_num((f_x * np.cos(phi) + f_y * np.sin(phi)) * np.cos(theta))
+
+        return DiffractionDataArray(
+            np.stack([f_theta, f_phi], axis=2), coords=self._make_coords_for_pol(["theta", "phi"])
+        )
+
     # pylint: disable=invalid-name
     @property
     def L_sph(self) -> DiffractionDataArray:
@@ -1338,6 +787,7 @@ class DiffractionData(MonitorData):
         """Radiation vectors associated with the magnetic field in spherical coordinates."""
         return self.car_2_sph(self.N)
 
+    # pylint: disable=invalid-name
     @property
     def E_sph(self) -> DiffractionDataArray:
         """Far field electric field in spherical coordinates, normalized so that
@@ -1355,6 +805,7 @@ class DiffractionData(MonitorData):
             np.stack([e_theta, e_phi], axis=2), coords=self._make_coords_for_pol(["theta", "phi"])
         )
 
+    # pylint: disable=invalid-name
     @property
     def H_sph(self) -> DiffractionDataArray:
         """Far field magnetic field in spherical coordinates."""
@@ -1377,26 +828,6 @@ class DiffractionData(MonitorData):
     def H_car(self) -> DiffractionDataArray:
         """Far field magnetic field in Cartesian coordinates."""
         return self.sph_2_car(self.H_sph)
-
-    @property
-    def amps(self) -> DiffractionDataArray:
-        """Complex power amplitude in each order for 's' and 'p' polarizations, normalized so that
-        the power carried by the wave of that order and polarization equals ``abs(amps)^2``.
-        """
-        cos_theta = np.cos(np.nan_to_num(self.angles[0]))
-        norm = 1.0 / np.sqrt(2.0 * self.eta) / np.sqrt(cos_theta)
-        amp_theta = self.E_sph.sel(polarization="theta").values * norm
-        amp_phi = self.E_sph.sel(polarization="phi").values * norm
-
-        # stack the amplitudes in s- and p-components along a new polarization axis
-        return DiffractionDataArray(
-            np.stack([amp_phi, amp_theta], axis=2), coords=self._make_coords_for_pol(["s", "p"])
-        )
-
-    @property
-    def power(self) -> xr.DataArray:
-        """Total power in each order, summed over both polarizations."""
-        return (np.abs(self.amps) ** 2).sum(dim="polarization")
 
 
 MonitorDataTypes = (
